requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
django==4.2.21
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
fsspec==2025.3.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
packaging==25.0
PyJWT==2.9.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sqlparse==0.5.3
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-extensions==4.13.2
urllib3==2.2.3
